refactor(app): separation of board code
