+++ src/enums/actionTypeEnum.ts
@@ -1,0 +1,6 @@
+export enum ActionTypeEnum {
+  DEFAULT = "none",
+  MOVING = "moving",
+  DRAWING = "drawing",
+  RESIZING = "resizing",
+}
