+++ src/enums/positionTypeEnum.ts
@@ -1,0 +1,6 @@
+export enum PositionTypeEnum {
+  DEFAULT = "none",
+  MOVING = "moving",
+  DRAWING = "drawing",
+  RESIZING = "resizing",
+}
