+++ src/interfaces/ElementsComponents.ts
@@ -1,0 +1,8 @@
+export default interface IElementsProps {
+  id: number;
+  xStart: number;
+  yStart: number;
+  xEnd: number;
+  yEnd: number;
+  type: string;
+}
