+++ src/types/actionType.ts
@@ -1,0 +1,1 @@
+export type ActionType = "none" | "moving" | "drawing" | "resizing";
