+++ src/types/positionType.ts
@@ -1,0 +1,1 @@
+export type PositionType = "tl" | "start" | "tr" | "bl" | "br" | "end";
