+++ src/views/Board/MouseUtils.ts
@@ -1,0 +1,14 @@
+export function cursorForPosition(position: string | null | undefined) {
+  switch (position) {
+    case "tl":
+    case "br":
+    case "start":
+    case "end":
+      return "nwse-resize";
+    case "tr":
+    case "bl":
+      return "nesw-resize";
+    default:
+      return "move";
+  }
+}
